day 17 faster and failproof
I wasn't considering the position in the jet pattern to search for
previous identical states. It was probably working because it's very
difficult to have 100 lines identical if the jet pattern wasn't the
same. However, this can fail in extreme cases, and the proof is that it
didn't work for my input if I checked only 20 lines instead of 100.

Now I check the jet pattern state too, I can reduce again to 20 lines
and have a faster execution, as well.